relocate message templates to utility file
