+++ lib/utils/sms-handler.js
@@ -1,18 +1,6 @@
 const client = require('twilio')(process.env.TWILIO_SID, process.env.TWILIO_TOKEN);
-
+const { onDemandMessage } = require('./message-templates');
 const User = require('../models/User');
-
-const dailyMessage = (user, location) =>
-  `
-    Here's your daily update from Social Distance-Ping!\nCOVID-19 stats for ${user.location} as of 4:30 pm PST today:\nNew Cases: ${location.newCases}\nNew Deaths: ${location.newDeaths}\nNew Recovered: ${location.newRecovered}`.trim();
-
-const onDemandMessage = stats => `
-Thanks for using Social Distance-Ping!
-COVID-19 stats for ${stats.location} as of ${stats.time} on ${stats.date}:
-New Cases: ${stats.newCases}
-New Deaths: ${stats.newDeaths}
-New Recovered: ${stats.newRecovered}
-`.trim();
 
 const outgoingSmsHandler = (to, message) => {
   return client.messages
@@ -24,6 +12,5 @@ const outgoingSmsHandler = (to, message) => {
 };
 
 module.exports = {
-  dailyMessage,
   outgoingSmsHandler
 };
